feat(dsg) move user-entity related files to a folder
